Neo 53889 adding abortable support (#74)
Currently there is no way to abort a request from the client. For example:

UI makes requests to fetch the data
While the call is in progress, the UI component unmounts/redirects to some other component.
While un-mounting, the UI should be able to cancel pending HTTP calls to avoid unnecessary attempts to re-render.